+++ src/kfs_frontend/kfs_grpc_inference_service.cpp
@@ -29,10 +29,15 @@
 #include "../deserialization.hpp"
 #include "../execution_context.hpp"
 #include "../grpc_utils.hpp"
-#include "../kfs_frontend/kfs_utils.hpp"
+#include "kfs_utils.hpp"
 #if (MEDIAPIPE_DISABLE == 0)
+// clang-format off
+// kfs_graph_executor_impl needs to be included before mediapipegraphexecutor
+// because it contains functions required by graph execution template
+#include "kfs_graph_executor_impl.hpp"
 #include "../mediapipe_internal/mediapipegraphdefinition.hpp"
 #include "../mediapipe_internal/mediapipegraphexecutor.hpp"
+// clang-format on
 #endif
 #include "../metric.hpp"
 #include "../modelinstance.hpp"
