+++ src/kfs_frontend/kfs_grpc_inference_service.hpp
@@ -31,6 +31,8 @@ using KFSModelMetadataRequest = inference::ModelMetadataRequest;
 using KFSModelMetadataResponse = inference::ModelMetadataResponse;
 using KFSRequest = inference::ModelInferRequest;
 using KFSResponse = inference::ModelInferResponse;
+using KFSStreamResponse = inference::ModelStreamInferResponse;
+using KFSServerReaderWriter = ::grpc::ServerReaderWriterInterface<KFSStreamResponse, KFSRequest>;
 using KFSTensorInputProto = inference::ModelInferRequest::InferInputTensor;
 using KFSTensorOutputProto = inference::ModelInferResponse::InferOutputTensor;
 using KFSShapeType = google::protobuf::RepeatedField<int64_t>;
